Fixed outdated information in badge.
By default, the checkout action will clone the repository at the
reference of the commit that triggered the action. Since the first job
pushes a commit to Git, the prior commit ref is used by the second job
when cloning the repository. By setting the ref to main, we ensure that
the most up-to-date commit has been cloned.