Improved percentage type detection for new table format.
BART updated their site and changed the format of some of the newer
ridership tables. Newer tables have a discrete <thead> instead of
shoving everything into a single <tbody>. This improves the detection
for percentage types in that case where the month(s) are no longer the
first row in the table body.