Replaced scraping script.
This commit completely replaces the scraping script used by this
repository to retrieve data from BART's website. Instead of using
a hand-rolled Python parser, this commit switches to a script built
on Flyscrape, a framework and runner for web scraping that automatically
produces JSON. BART has normalized the format in which they present
their data, so the detection logic employed by the prior Python version
is no longer required. This should make maintenance a bit easier
for myself.

- Added `scripts/ridership.js`. This script is run using Flyscrape.
- Updated 'scheduled' workflow to use the flyscrape action I wrote
  to run the aforementioned script. The resulting scraped content
  is then merged with the existing ridership data archive.
- Added step to install moreutils package so I can use `sponge`.
- Removed old collector script and related dependency management
  files.